requirements.txt
langchain
chromadb
python_dotenv